Update workload for quality and deployment (#35)
* minor updates

* fixed bug that removed all reactions and renamed non premium channels and minor updates on ready guild channel check

* implemented delete  channel and delete message event updates

* fixed an error where user portal creator could not update portals data when in voice that wasnt his and auto deletion for both portal's and users message

* updated workflow and ranks to not break

+++ functions/help_manager.js
@@ -68,18 +68,19 @@ module.exports = {
 			false,
 			yts.thumbnail,
 		);
-		const channel = guild_object.channels.cache
-			.get(guild.music_data.channel_id);
+		const channel = guild_object.channels.cache.get(guild.music_data.channel_id);
 
-		channel.messages.channel.messages
-			.fetch(guild.music_data.message_id)
-			.then(message => {
-				message.edit(music_message_emb)
-					.then(msg =>
-						console.log(`Updated the content of a message to ${msg.content}`))
-					.catch(console.error);
-			})
-			.catch(console.error);
+		if(channel) {
+			channel.messages.channel.messages
+				.fetch(guild.music_data.message_id)
+				.then(message => {
+					message.edit(music_message_emb)
+						.then(msg =>
+							console.log(`Updated the content of a message to ${msg.content}`))
+						.catch(console.error);
+				})
+				.catch(console.error);
+		}
 	},
 
 	join_user_voice: async function(client, message, portal_guilds, join) { // localize
@@ -215,21 +216,28 @@ module.exports = {
 
 	empty_channel_remover: function(current_guild, portal_guilds, portal_managed_guilds_path) {
 		current_guild.channels.cache.forEach(channel => {
-			for(const portal_channel in portal_guilds[current_guild.id].portal_list) {
-				if(portal_guilds[current_guild.id].portal_list[portal_channel].voice_list[channel.id]) {
-					if(!channel.members.size) {
-						console.log('Deleting channel: ', channel.name, 'from ', channel.guild.name);
-						// guld_mngr.delete_channel(channel);
-						if (channel.deletable) {
-							channel
-								.delete()
-								.then(g => console.log(`Deleted channel with id: ${g}`))
-								.catch(console.error);
+			if(portal_guilds[current_guild.id]) {
+				for(const portal_channel in portal_guilds[current_guild.id].portal_list) {
+					if(portal_guilds[current_guild.id].portal_list[portal_channel].voice_list[channel.id]) {
+						if(!channel.members.size) {
+							console.log('Deleting channel: ', channel.name, 'from ', channel.guild.name);
+							// guld_mngr.delete_channel(channel);
+							if (channel.deletable) {
+								channel
+									.delete()
+									.then(g => console.log(`Deleted channel with id: ${g}`))
+									.catch(console.error);
+							}
+							return true;
 						}
-						return true;
 					}
+					return false;
 				}
-				return false;
+			}
+			else {
+				current_guild.leave()
+					.then(guild => console.log(`Left guild ${guild}`))
+					.catch(console.error);
 			}
 		});
 
@@ -283,10 +291,12 @@ module.exports = {
 			}
 			else if (status === true) {
 				message.react('✔️');
+				message.delete({ timeout: 5000 });
 			}
 			else if (status === false) {
 				lclz_mngr.client_talk(client, portal_guilds, 'fail');
 				message.react('❌');
+				message.delete({ timeout: 5000 });
 			}
 		}
 	},
